predict stars by review
I'm using the yelp data set to predict the stars of single restaurant. 
First, I combined two files which contains business_id, stars and reviews relatively together. As the original file is far too large, so I get around 15000 different restaurants and all corresponding reviews. For convenience, I output the file with extension .csv. Then I used stopwords module to remove the useless stop words in the reviews and get the vector of bags of words. Then I applied random forest to predict the stars with the bags of words. The final accuracy is around 48%